requirements.txt
pandas
openpyxl
google-generativeai
Pillow
flask
python-dotenv
